Add files via upload
If you run this inside an file folder the code will converte the file with '.mp4' extension into an '.mp3'.
When i was using this code i notice that moviepy library there is a loss of 8 to 10 seconds of audio in the final file. I'm still searching about this problem.